prevent spurious changes in debian-usptream-obsolete-path.

+++ fixers/debian-upstream-obsolete-path.sh
@@ -2,8 +2,10 @@
 if [[ -f debian/upstream ]]; then
 	mv debian/upstream debian/upstream-metadata.yaml
 fi
-mkdir -p debian/upstream
-test -f debian/upstream-metadata && mv debian/upstream-metadata debian/upstream/metadata
-test -f debian/upstream-metadata.yaml && mv debian/upstream-metadata.yaml debian/upstream/metadata
+if test -f debian/upstream-metadata -o -f debian/upstream-metadata.yaml; then
+	mkdir -p debian/upstream
+	test -f debian/upstream-metadata && mv debian/upstream-metadata debian/upstream/metadata
+	test -f debian/upstream-metadata.yaml && mv debian/upstream-metadata.yaml debian/upstream/metadata
+fi
 echo "Move upstream metadata to debian/upstream/metadata."
 echo "Fixed-Lintian-Tags: debian-upstream-obsolete-path"
